Use a different approach to parse the incoming bytes
This will clear out the buffer if there are wrong bytes along the way and only keep one packet read in at a time.

It also removed the need for member variables to store the frame information as that can be stored locally and passed onto the final processing function.